requirements.txt
Django==4.2.10
scapy==2.5.0
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
tensorflow==2.15.0
matplotlib==3.7.2
reportlab==4.0.7
requests==2.31.0
psycopg2-binary==2.9.9
gunicorn==21.2.0
pyyaml==6.0.1
python-dotenv==1.0.0
djangorestframework==3.14.0
pycryptodome==3.19.0
tqdm==4.66.1
opencv-python==4.8.1.78
plotly==5.18.0
nltk==3.8.1
pillow==10.1.0
pytest==7.4.3
django-crispy-forms==2.0
crispy-bootstrap5==0.7

# Added security dependencies
cryptography==41.0.3
PyJWT==2.8.0
bcrypt==4.0.1
passlib==1.7.4
argon2-cffi==23.1.0

# Advanced ML dependencies
xgboost==2.0.0
lightgbm==4.1.0
torch==2.0.1
transformers==4.35.2
sentence-transformers==2.2.2
shap==0.43.0
lime==0.2.0.1
captum==0.6.0
alibi==0.9.4

# Device support and cross-platform
pywin32==306; sys_platform == 'win32'
paramiko==3.3.1
netmiko==4.2.0
pysnmp==4.4.12
netifaces==0.11.0
pypcap==1.3.0
dnspython==2.4.2
pyopenssl==23.3.0

# Cloud integrations
boto3==1.28.50
azure-identity==1.14.1
azure-mgmt-compute==30.3.0
google-cloud-compute==1.14.1
google-cloud-storage==2.10.0

# IoT support
paho-mqtt==2.2.1
coap==0.2.0
iotzpy==1.1.0
zeroconf==0.115.0

# Industrial and SCADA support
pymodbus==3.5.2
pycomm3==1.2.8
modbus-tk==1.1.3

# API and web security
django-cors-headers==4.3.0
django-oauth-toolkit==2.3.0
drf-jwt==1.19.2
bleach==6.1.0
django-csp==3.7
celery==5.3.4
redis==5.0.1

# Visualization and reporting
dash==2.13.0
bokeh==3.2.2
seaborn==0.13.0
django-q==1.3.9
dask==2023.10.1
django-prometheus==2.3.1 